+++ packages/components-react/language-navigation-react/src/index.ts
@@ -1,3 +1,6 @@
+import '@rijkshuisstijl-community/icon-react';
 import '@rijkshuisstijl-community/language-navigation-css';
+import '@rijkshuisstijl-community/link-button-react';
+import '@rijkshuisstijl-community/link-react';
 
 export * from './LanguageNavigation';