+++ packages/components-react/link-list-card-react/src/index.ts
@@ -1,3 +1,5 @@
+import '@rijkshuisstijl-community/heading-react';
 import '@rijkshuisstijl-community/link-list-card-css';
+import '@rijkshuisstijl-community/link-list-react';
 
 export * from './LinkListCard';