+++ packages/components-react/logo-react/src/Logo.tsx
@@ -4,12 +4,14 @@
  */
 
 import clsx from 'clsx';
-import { HTMLAttributes, PropsWithChildren, ReactNode, Ref } from 'react';
+import { AnchorHTMLAttributes, HTMLAttributes, PropsWithChildren, ReactNode, Ref } from 'react';
 
 export interface LogoProps extends HTMLAttributes<HTMLDivElement> {
-  organisation: ReactNode;
+  organisation?: ReactNode;
   subtitle?: ReactNode;
   ref?: Ref<HTMLDivElement>;
+  href?: string;
+  hrefProps?: AnchorHTMLAttributes<HTMLAnchorElement>;
 }
 
 export const Logo = ({
@@ -18,15 +20,31 @@ export const Logo = ({
   className,
   organisation,
   subtitle,
+  href,
+  hrefProps,
   ...restProps
 }: PropsWithChildren<LogoProps>) => {
+  const content = (
+    <>
+      <div className="rhc-logo__image">{children}</div>
+      {organisation ? (
+        <figcaption className="rhc-logo__caption">
+          <p className="rhc-logo__title">{organisation}</p>
+          {subtitle && <p className="rhc-logo__subtitle">{subtitle}</p>}
+        </figcaption>
+      ) : null}
+    </>
+  );
+
   return (
     <figure className={clsx('rhc-logo', className)} ref={ref} {...restProps}>
-      <div className="rhc-logo__image">{children}</div>
-      <figcaption className="rhc-logo__caption">
-        <p className="rhc-logo__title">{organisation}</p>
-        {subtitle && <p className="rhc-logo__subtitle">{subtitle}</p>}
-      </figcaption>
+      {href ? (
+        <a {...hrefProps} className={clsx('rhc-logo__link', hrefProps?.className)} href={href}>
+          {content}
+        </a>
+      ) : (
+        content
+      )}
     </figure>
   );
 };