+++ packages/components-react/navigation-bar-react/src/index.ts
@@ -1,3 +1,8 @@
+import '@rijkshuisstijl-community/column-layout-react';
+import '@rijkshuisstijl-community/heading-react';
+import '@rijkshuisstijl-community/icon-react';
+import '@rijkshuisstijl-community/link-list-react';
+import '@rijkshuisstijl-community/link-react';
 import '@rijkshuisstijl-community/navigation-bar-css';
 
 export * from './NavBar';