+++ packages/components-react/side-nav-link-react/src/index.ts
@@ -1,0 +1,3 @@
+import '@rijkshuisstijl-community/icon-react';
+
+export * from './SideNavLink';