+++ packages/components-react/toggletip-react/src/index.ts
@@ -1,3 +1,6 @@
+import '@rijkshuisstijl-community/alert-react';
+import '@rijkshuisstijl-community/icon-react';
+import '@rijkshuisstijl-community/paragraph-react';
 import '@rijkshuisstijl-community/toggletip-css';
 
 export * from './Toggletip';