+++ packages/storybook/src/components-react/separator.stories.tsx
@@ -20,22 +20,9 @@ const meta = {
     github:
       'https://github.com/nl-design-system/rijkshuisstijl-community/blob/main/packages/components-react/src/Separator.tsx',
   },
-  argTypes: {
-    invisible: {
-      control: {
-        type: 'boolean',
-      },
-    },
-  },
 } satisfies Meta<typeof Separator>;
 
 export default meta;
 
 type Story = StoryObj<typeof meta>;
 export const Default: Story = {};
-
-export const Invisible: Story = {
-  args: {
-    invisible: true,
-  },
-};