+++ packages/components-react/accordion-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rijkshuisstijl-community/accordion-react
 
+## 3.1.1
+
+### Patch Changes
+
+- e41c59f: update dependencies
+
 ## 3.1.0
 
 ### Minor Changes