+++ packages/components-react/icon-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rijkshuisstijl-community/icon-react
 
+## 3.1.2
+
+### Patch Changes
+
+- b3db8e1: Icoon chevron-left toegevoegd
+
+## 3.1.1
+
+### Patch Changes
+
+- e41c59f: update dependencies
+
 ## 3.1.0
 
 ### Minor Changes