+++ packages/components-react/listbox-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rijkshuisstijl-community/listbox-react
 
+## 1.0.1
+
+### Patch Changes
+
+- e41c59f: update dependencies
+- Updated dependencies [e41c59f]
+  - @rijkshuisstijl-community/listbox-css@1.0.1
+
 ## 1.0.0
 
 ### Major Changes